requirements.txt
keras==2.8.0
Kivy==2.0.0
layers==0.1.5
numpy==1.22.2
opencv_python==4.5.5.62
Pillow==9.0.1
tensorflow==2.8.0
